add element service class

+++ share-basic/share-dao/src/main/java/org/andot/share/basic/entity/MenuElement.java
@@ -3,10 +3,15 @@ package org.andot.share.basic.entity;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
+/**
+ * 菜单的元素表
+ * @author andot
+ */
 @Data
 @EqualsAndHashCode(callSuper = false)
 public class MenuElement extends BaseEntity {
-    private Integer menuElementId;
-    private Integer menuId;
-    private Integer elementId;
+    private Long menuElementId;
+    private Long menuId;
+    private Long elementId;
+
 }
