+++ share-basic/share-dao/src/main/java/org/andot/share/basic/entity/PageElement.java
@@ -3,10 +3,14 @@ package org.andot.share.basic.entity;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
+/**
+ *
+ * @author andot
+ */
 @Data
 @EqualsAndHashCode(callSuper = false)
 public class PageElement extends BaseEntity {
-    private Integer pageElementId;
+    private Long pageElementId;
     private String pageElementName;
     private String pageElementCode;
     private Byte pageElementType;
